Fix use of Throttle for ebusd update method
The way ebusd component is implemented it makes it impossible
to guarantee that update of individual sensors won't be rejected.
This happens because EbusdData class handles all read requests from
all sensors and is decorated with @Throttle. Thus, it is possible
that sequential read requests from sensors will not be handled and
make sort of race condition making sensor's update not deterministic.

Workaround this by moving @Throttle to EbusdSensor class' update
method. This way we protect individual sensor from being queried
too frequently and give equal opportunities to all entities.