Do not raise RuntimeError if value can't be converted
In case some of the monitored conditions return value which
cannot be converted it is not expected to raise RuntimeError
as the value can get back to normal, e.g. if heating switched
off then room temperature can be returned as '-'. Raising
RuntimeError will exclude the monitored condition (sensor) from
further polling which is not desired.